requirements.txt
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
opencv-python>=4.8.0
pillow>=10.0.0
openai>=1.3.0
python-dotenv>=1.0.0
dlib>=19.24.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
mediapipe>=0.10.0
gradio>=4.8.0
xformers>=0.0.22
numpy>=1.24.0